+++ llvm/include/llvm/IR/InstrTypes.h
@@ -25,6 +25,7 @@
 #include "llvm/IR/CallingConv.h"
 #include "llvm/IR/DerivedTypes.h"
 #include "llvm/IR/FMF.h"
+#include "llvm/IR/FPEnv.h"
 #include "llvm/IR/Function.h"
 #include "llvm/IR/Instruction.h"
 #include "llvm/IR/LLVMContext.h"
@@ -1097,6 +1098,28 @@ public:
 using OperandBundleDef = OperandBundleDefT<Value *>;
 using ConstOperandBundleDef = OperandBundleDefT<const Value *>;
 
+std::optional<StringRef> getBundleOperandByPrefix(OperandBundleUse Bundle,
+                                                  StringRef Prefix);
+void addOperandToBundleTag(LLVMContext &Ctx,
+                           SmallVectorImpl<OperandBundleDef> &Bundles,
+                           StringRef Tag, size_t PrefixSize, StringRef Val);
+
+void addFPRoundingBundle(LLVMContext &Ctx,
+                         SmallVectorImpl<OperandBundleDef> &Bundles,
+                         RoundingMode Rounding);
+void addFPExceptionBundle(LLVMContext &Ctx,
+                          SmallVectorImpl<OperandBundleDef> &Bundles,
+                          fp::ExceptionBehavior Except);
+std::optional<DenormalMode::DenormalModeKind>
+getDenormModeBundle(const OperandBundleUse &Control, bool Unput,
+                    const fltSemantics *FPSem);
+void addInputDenormBundle(LLVMContext &Ctx,
+                          SmallVectorImpl<OperandBundleDef> &Bundles,
+                          DenormalMode::DenormalModeKind Mode);
+void addOutputDenormBundle(LLVMContext &Ctx,
+                           SmallVectorImpl<OperandBundleDef> &Bundles,
+                           DenormalMode::DenormalModeKind Mode);
+
 //===----------------------------------------------------------------------===//
 //                               CallBase Class
 //===----------------------------------------------------------------------===//
@@ -1155,6 +1178,8 @@ protected:
   /// Get the number of extra operands for instructions that don't have a fixed
   /// number of extra operands.
   LLVM_ABI unsigned getNumSubclassExtraOperandsDynamic() const;
+
+  MemoryEffects getFloatingPointMemoryEffects() const;
 
 public:
   using Instruction::getContext;
@@ -2165,6 +2190,28 @@ public:
     }
     return false;
   }
+
+  /// Return rounding mode specified for this call.
+  RoundingMode getRoundingMode() const;
+
+  /// Return exception behavior specified for this call.
+  fp::ExceptionBehavior getExceptionBehavior() const;
+
+  /// Return input denormal mode specified by operand bundles.
+  std::optional<DenormalMode::DenormalModeKind>
+  getInputDenormMode(const fltSemantics *FPSem = nullptr) const;
+
+  /// Return output denormal mode specified by operand bundles.
+  std::optional<DenormalMode::DenormalModeKind>
+  getOutputDenormMode(const fltSemantics *FPSem = nullptr) const;
+
+  /// Return input denormal mode specified by operand bundles.
+  std::optional<DenormalMode::DenormalModeKind>
+  getInputDenormModeFromBundle(const fltSemantics *FPSem = nullptr) const;
+
+  /// Return output denormal mode specified by operand bundles.
+  std::optional<DenormalMode::DenormalModeKind>
+  getOutputDenormModeFromBundle(const fltSemantics *FPSem = nullptr) const;
 
   /// Used to keep track of an operand bundle.  See the main comment on
   /// OperandBundleUser above.