+++ llvm/lib/Transforms/Utils/Local.cpp
@@ -503,10 +503,10 @@ bool llvm::wouldInstructionBeTriviallyDead(const Instruction *I,
       return false;
     }
 
-    if (auto *FPI = dyn_cast<ConstrainedFPIntrinsic>(I)) {
-      std::optional<fp::ExceptionBehavior> ExBehavior =
-          FPI->getExceptionBehavior();
-      return *ExBehavior != fp::ebStrict;
+    if (auto *II = dyn_cast<IntrinsicInst>(I);
+        II && Intrinsic::isConstrainedFPIntrinsic(II->getIntrinsicID())) {
+      // llvm.fcmps uses the fp.except operand bundle for exception behavior.
+      return II->getExceptionBehavior() != fp::ebStrict;
     }
   }